pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Artur Araujo <arturavs1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pydentic = "^0.0.1.dev3"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
pydantic-settings = "^2.6.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[tool.ruff]
extend-exlude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I','F','E','W','PL','PT']

[tool.ruff.format]
preview = true
quote-style = 'single'



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
